fix(jobs): share sandbox terminal client
