+++ ui/src/components/runner-job-detail.tsx
@@ -1,8 +1,5 @@
-import { type ReactNode, useCallback, useEffect, useMemo, useRef, useState } from "react"
+import { type ReactNode, useCallback, useEffect, useMemo, useState } from "react"
 import { ArrowLeft, CheckCircle2, Clock3, ExternalLink, Loader2, Play, RefreshCw, SquareTerminal, XCircle } from "lucide-react"
-import { Terminal } from "xterm"
-import { FitAddon } from "xterm-addon-fit"
-import "xterm/css/xterm.css"
 
 import type { RunnerJobGroup, RunnerState } from "@/admin-types"
 import { logNames } from "@/admin-types"
@@ -11,6 +8,7 @@ import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import { useSandboxTerminal } from "@/hooks/use-sandbox-terminal"
 import { cn } from "@/lib/utils"
 
 type LogName = (typeof logNames)[number]
@@ -23,12 +21,6 @@ type RunnerJobDetailProps = {
   request: (url: string, options?: RequestInit) => Promise<unknown>
 }
 
-type TerminalCreateResponse = {
-  session_id: string
-  pid: number
-  sandbox_id: string
-}
-
 export function RunnerJobDetail({ id, apiBase, onBack, onOpenJob, request }: RunnerJobDetailProps) {
   const [job, setJob] = useState<RunnerState | null>(null)
   const [jobGroup, setJobGroup] = useState<RunnerJobGroup | null>(null)
@@ -37,19 +29,16 @@ export function RunnerJobDetail({ id, apiBase, onBack, onOpenJob, request }: Run
   const [logText, setLogText] = useState("Loading log...")
   const [githubLogText, setGithubLogText] = useState("Loading GitHub log...")
   const [githubLogLoading, setGithubLogLoading] = useState(false)
-  const [terminalSession, setTerminalSession] = useState<TerminalCreateResponse | null>(null)
-  const [terminalError, setTerminalError] = useState("")
-  const [terminalConnecting, setTerminalConnecting] = useState(false)
-  const terminalEl = useRef<HTMLDivElement | null>(null)
-  const terminalRef = useRef<Terminal | null>(null)
-  const fitRef = useRef<FitAddon | null>(null)
-  const eventSourceRef = useRef<EventSource | null>(null)
-  const terminalDataDisposableRef = useRef<{ dispose: () => void } | null>(null)
-  const terminalSessionRef = useRef<TerminalCreateResponse | null>(null)
-  const resizeTimeoutRef = useRef<number | null>(null)
-
   const endpoint = useMemo(() => `${apiBase}/${encodeURIComponent(id)}`, [apiBase, id])
   const terminalAvailable = job ? isTerminalAvailable(job) : false
+  const { terminalEl, terminalSession, terminalError, terminalConnecting, connectTerminal } = useSandboxTerminal({
+    endpoint,
+    available: terminalAvailable,
+    request,
+    connectingMessage: "Connecting to sandbox terminal...",
+    streamDisconnectedMessage: "Terminal stream disconnected",
+    connectErrorMessage: "Failed to connect terminal",
+  })
 
   const loadJob = useCallback(async () => {
     setLoading(true)
@@ -100,40 +89,6 @@ export function RunnerJobDetail({ id, apiBase, onBack, onOpenJob, request }: Run
     void loadGithubLog()
   }, [loadGithubLog])
 
-  const closeTerminalSession = useCallback(
-    (session = terminalSessionRef.current) => {
-      eventSourceRef.current?.close()
-      eventSourceRef.current = null
-      terminalDataDisposableRef.current?.dispose()
-      terminalDataDisposableRef.current = null
-      if (resizeTimeoutRef.current) {
-        window.clearTimeout(resizeTimeoutRef.current)
-        resizeTimeoutRef.current = null
-      }
-      terminalRef.current?.dispose()
-      terminalRef.current = null
-      fitRef.current = null
-      terminalSessionRef.current = null
-      setTerminalSession(null)
-      if (session) {
-        void fetch(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}`, {
-          method: "DELETE",
-          credentials: "same-origin",
-          keepalive: true,
-        }).catch(() => undefined)
-      }
-    },
-    [endpoint],
-  )
-
-  useEffect(() => {
-    setTerminalError("")
-    closeTerminalSession()
-    return () => {
-      closeTerminalSession()
-    }
-  }, [closeTerminalSession])
-
   const refreshPage = () => {
     void loadJob()
     void loadJobGroup()
@@ -147,133 +102,6 @@ export function RunnerJobDetail({ id, apiBase, onBack, onOpenJob, request }: Run
     }
     window.history.pushState(null, "", `/jobs/${encodeURIComponent(jobID)}`)
   }
-
-  const connectTerminal = async () => {
-    if (!terminalAvailable || terminalConnecting || terminalSession) return
-    setTerminalConnecting(true)
-    setTerminalError("")
-    try {
-      const term = new Terminal({
-        cursorBlink: true,
-        convertEol: true,
-        fontFamily: "var(--font-mono)",
-        fontSize: 13,
-        theme: {
-          background: "#111318",
-          foreground: "#e6edf3",
-          cursor: "#36d399",
-          selectionBackground: "#334155",
-        },
-      })
-      const fit = new FitAddon()
-      term.loadAddon(fit)
-      terminalRef.current = term
-      fitRef.current = fit
-      if (terminalEl.current) {
-        term.open(terminalEl.current)
-        fit.fit()
-      }
-      term.writeln("Connecting to sandbox terminal...")
-      const cols = term.cols || 100
-      const rows = term.rows || 28
-      const session = (await request(`${endpoint}/terminal`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ cols, rows }),
-      })) as TerminalCreateResponse
-      if (terminalRef.current !== term) {
-        void fetch(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}`, {
-          method: "DELETE",
-          credentials: "same-origin",
-          keepalive: true,
-        }).catch(() => undefined)
-        return
-      }
-      setTerminalSession(session)
-      terminalSessionRef.current = session
-      term.writeln(`Connected to ${session.sandbox_id} pid=${session.pid}`)
-      const events = new EventSource(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/events`, {
-        withCredentials: true,
-      })
-      eventSourceRef.current = events
-      events.onmessage = (event) => {
-        try {
-          term.write(JSON.parse(event.data) as string)
-        } catch {
-          term.write(event.data)
-        }
-      }
-      events.onerror = () => {
-        setTerminalError("Terminal stream disconnected")
-        closeTerminalSession(session)
-      }
-      let inputClosed = false
-      let inputSending = false
-      const inputQueue: string[] = []
-      const sendNextInput = async () => {
-        if (inputClosed || inputSending || inputQueue.length === 0) return
-        inputSending = true
-        const data = inputQueue.join("")
-        inputQueue.length = 0
-        try {
-          await request(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/input`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ data }),
-          })
-        } catch (error) {
-          if (!inputClosed) {
-            setTerminalError(error instanceof Error ? error.message : "Failed to send input")
-          }
-        } finally {
-          inputSending = false
-          void sendNextInput()
-        }
-      }
-      const inputDisposable = term.onData((data) => {
-        inputQueue.push(data)
-        void sendNextInput()
-      })
-      terminalDataDisposableRef.current = {
-        dispose: () => {
-          inputClosed = true
-          inputQueue.length = 0
-          inputDisposable.dispose()
-        },
-      }
-    } catch (error) {
-      setTerminalError(error instanceof Error ? error.message : "Failed to connect terminal")
-      closeTerminalSession()
-    } finally {
-      setTerminalConnecting(false)
-    }
-  }
-
-  const resizeTerminal = useCallback(() => {
-    const session = terminalSession
-    const term = terminalRef.current
-    const fit = fitRef.current
-    if (!session || !term || !fit) return
-    fit.fit()
-    if (resizeTimeoutRef.current) {
-      window.clearTimeout(resizeTimeoutRef.current)
-    }
-    resizeTimeoutRef.current = window.setTimeout(() => {
-      resizeTimeoutRef.current = null
-      void request(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/resize`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ cols: term.cols, rows: term.rows }),
-      }).catch(() => undefined)
-    }, 250)
-  }, [endpoint, request, terminalSession])
-
-  useEffect(() => {
-    if (!terminalSession) return
-    const observer = new ResizeObserver(resizeTerminal)
-    if (terminalEl.current) observer.observe(terminalEl.current)
-    return () => observer.disconnect()
-  }, [terminalSession, endpoint, resizeTerminal])
 
   const groupJobs = jobGroup ? [...jobGroup.current_jobs, ...jobGroup.previous_jobs] : job ? [job] : []
