+++ ui/src/components/user-dashboard.tsx
@@ -19,10 +19,7 @@ import {
   X,
 } from "lucide-react"
 import { useTheme } from "next-themes"
-import { type CSSProperties, type FormEvent, type MouseEvent, type ReactNode, useCallback, useEffect, useMemo, useRef, useState } from "react"
-import { Terminal } from "xterm"
-import { FitAddon } from "xterm-addon-fit"
-import "xterm/css/xterm.css"
+import { type CSSProperties, type FormEvent, type MouseEvent, type ReactNode, useEffect, useMemo, useRef, useState } from "react"
 
 import type { AuthSession, GitHubAppConfig, RunnerJobGroup, RunnerState, UserPreferences } from "@/admin-types"
 import { logNames } from "@/admin-types"
@@ -43,6 +40,7 @@ import {
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import { useSandboxTerminal } from "@/hooks/use-sandbox-terminal"
 import { cn } from "@/lib/utils"
 
 type BuildGroupKind = "pull_request" | "branch" | "workflow_run" | "manual"
@@ -67,12 +65,6 @@ type AccountSettingsTab = "repositories" | "preferences"
 type AccountSettingsRoute = {
   accountLogin?: string
   tab: AccountSettingsTab
-}
-
-type TerminalCreateResponse = {
-  session_id: string
-  pid: number
-  sandbox_id: string
 }
 
 type GitHubLogState =
@@ -1031,61 +1023,29 @@ function RunnerJobLogPanel({
   const [runnerLogText, setRunnerLogText] = useState("Loading runner log...")
   const [githubLog, setGithubLog] = useState<GitHubLogState>({ kind: "log", text: "Loading GitHub log..." })
   const [githubLogLoading, setGithubLogLoading] = useState(false)
-  const [terminalSession, setTerminalSession] = useState<TerminalCreateResponse | null>(null)
-  const [terminalError, setTerminalError] = useState("")
-  const [terminalConnecting, setTerminalConnecting] = useState(false)
-  const terminalEl = useRef<HTMLDivElement | null>(null)
-  const terminalRef = useRef<Terminal | null>(null)
-  const fitRef = useRef<FitAddon | null>(null)
-  const eventSourceRef = useRef<EventSource | null>(null)
-  const terminalDataDisposableRef = useRef<{ dispose: () => void } | null>(null)
-  const terminalSessionRef = useRef<TerminalCreateResponse | null>(null)
-  const resizeTimeoutRef = useRef<number | null>(null)
   const endpoint = `/user/runner_requests/${encodeURIComponent(job.id)}`
   const endpointRef = useRef(endpoint)
   const terminalAvailable = isTerminalAvailable(job)
+  const { terminalEl, terminalSession, terminalError, terminalConnecting, connectTerminal } = useSandboxTerminal({
+    endpoint,
+    available: terminalAvailable,
+    request,
+    connectingMessage: "Connecting to sandbox web console...",
+    streamDisconnectedMessage: "Web console stream disconnected",
+    connectErrorMessage: "Failed to connect web console",
+  })
 
   useEffect(() => {
     endpointRef.current = endpoint
   }, [endpoint])
 
-  const closeTerminalSession = useCallback(
-    (session = terminalSessionRef.current) => {
-      eventSourceRef.current?.close()
-      eventSourceRef.current = null
-      terminalDataDisposableRef.current?.dispose()
-      terminalDataDisposableRef.current = null
-      if (resizeTimeoutRef.current) {
-        window.clearTimeout(resizeTimeoutRef.current)
-        resizeTimeoutRef.current = null
-      }
-      terminalRef.current?.dispose()
-      terminalRef.current = null
-      fitRef.current = null
-      terminalSessionRef.current = null
-      setTerminalSession(null)
-      if (session) {
-        void fetch(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}`, {
-          method: "DELETE",
-          credentials: "same-origin",
-          keepalive: true,
-        }).catch(() => undefined)
-      }
-    },
-    [endpoint],
-  )
-
-  useEffect(() => {
-    setTerminalError("")
-    closeTerminalSession()
-    return () => {
-      closeTerminalSession()
-    }
-  }, [closeTerminalSession])
-
   useEffect(() => {
     let active = true
-    setRunnerLogText("Loading runner log...")
+    queueMicrotask(() => {
+      if (active) {
+        setRunnerLogText("Loading runner log...")
+      }
+    })
     void request(`${endpoint}/logs/${encodeURIComponent(selectedLog)}`)
       .then((text) => {
         if (active) {
@@ -1104,8 +1064,12 @@ function RunnerJobLogPanel({
 
   useEffect(() => {
     let active = true
-    setGithubLogLoading(true)
-    setGithubLog({ kind: "log", text: "Loading GitHub log..." })
+    queueMicrotask(() => {
+      if (active) {
+        setGithubLogLoading(true)
+        setGithubLog({ kind: "log", text: "Loading GitHub log..." })
+      }
+    })
     void request(`${endpoint}/github-log`)
       .then((text) => {
         if (active) {
@@ -1148,133 +1112,6 @@ function RunnerJobLogPanel({
         }
       })
   }
-
-  const connectTerminal = async () => {
-    if (!terminalAvailable || terminalConnecting || terminalSession) return
-    setTerminalConnecting(true)
-    setTerminalError("")
-    try {
-      const term = new Terminal({
-        cursorBlink: true,
-        convertEol: true,
-        fontFamily: "var(--font-mono)",
-        fontSize: 13,
-        theme: {
-          background: "#111318",
-          foreground: "#e6edf3",
-          cursor: "#36d399",
-          selectionBackground: "#334155",
-        },
-      })
-      const fit = new FitAddon()
-      term.loadAddon(fit)
-      terminalRef.current = term
-      fitRef.current = fit
-      if (terminalEl.current) {
-        term.open(terminalEl.current)
-        fit.fit()
-      }
-      term.writeln("Connecting to sandbox web console...")
-      const cols = term.cols || 100
-      const rows = term.rows || 28
-      const session = (await request(`${endpoint}/terminal`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ cols, rows }),
-      })) as TerminalCreateResponse
-      if (terminalRef.current !== term) {
-        void fetch(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}`, {
-          method: "DELETE",
-          credentials: "same-origin",
-          keepalive: true,
-        }).catch(() => undefined)
-        return
-      }
-      setTerminalSession(session)
-      terminalSessionRef.current = session
-      term.writeln(`Connected to ${session.sandbox_id} pid=${session.pid}`)
-      const events = new EventSource(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/events`, {
-        withCredentials: true,
-      })
-      eventSourceRef.current = events
-      events.onmessage = (event) => {
-        try {
-          term.write(JSON.parse(event.data) as string)
-        } catch {
-          term.write(event.data)
-        }
-      }
-      events.onerror = () => {
-        setTerminalError("Web console stream disconnected")
-        closeTerminalSession(session)
-      }
-      let inputClosed = false
-      let inputSending = false
-      const inputQueue: string[] = []
-      const sendNextInput = async () => {
-        if (inputClosed || inputSending || inputQueue.length === 0) return
-        inputSending = true
-        const data = inputQueue.join("")
-        inputQueue.length = 0
-        try {
-          await request(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/input`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ data }),
-          })
-        } catch (error) {
-          if (!inputClosed) {
-            setTerminalError(error instanceof Error ? error.message : "Failed to send input")
-          }
-        } finally {
-          inputSending = false
-          void sendNextInput()
-        }
-      }
-      const inputDisposable = term.onData((data) => {
-        inputQueue.push(data)
-        void sendNextInput()
-      })
-      terminalDataDisposableRef.current = {
-        dispose: () => {
-          inputClosed = true
-          inputQueue.length = 0
-          inputDisposable.dispose()
-        },
-      }
-    } catch (error) {
-      setTerminalError(error instanceof Error ? error.message : "Failed to connect web console")
-      closeTerminalSession()
-    } finally {
-      setTerminalConnecting(false)
-    }
-  }
-
-  const resizeTerminal = useCallback(() => {
-    const session = terminalSession
-    const term = terminalRef.current
-    const fit = fitRef.current
-    if (!session || !term || !fit) return
-    fit.fit()
-    if (resizeTimeoutRef.current) {
-      window.clearTimeout(resizeTimeoutRef.current)
-    }
-    resizeTimeoutRef.current = window.setTimeout(() => {
-      resizeTimeoutRef.current = null
-      void request(`${endpoint}/terminal/${encodeURIComponent(session.session_id)}/resize`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ cols: term.cols, rows: term.rows }),
-      }).catch(() => undefined)
-    }, 250)
-  }, [endpoint, request, terminalSession])
-
-  useEffect(() => {
-    if (!terminalSession) return
-    const observer = new ResizeObserver(resizeTerminal)
-    if (terminalEl.current) observer.observe(terminalEl.current)
-    return () => observer.disconnect()
-  }, [terminalSession, resizeTerminal])
 
   const githubLogActions = (
     <Button
